fix(boxPlot): Wide scatter for constant y
When kernelDensity interpolation fails, it is because all values in y are the same. In such cases, it is desired to have maximum possible horizontal spread of samples in the scatter (as it would have if all but one value in y are the same, or values in y are close together). Thus, maxDisplacement should be 1, not 0, in the catch clause.

Fixes IoSR-Surrey/MatlabToolbox#14